+++ src/main/java/top/vuhe/controller/factory/Factory.java
@@ -1,0 +1,20 @@
+package top.vuhe.controller.factory;
+
+import top.vuhe.model.entity.Question;
+
+/**
+ * @author vuhe
+ */
+public abstract class Factory<T> {
+
+    public static Factory<Question> getQuestionFactory() {
+        return new QuestionFactory();
+    }
+
+    /**
+     * 生产方法
+     * @return 生产的对象
+     */
+    abstract public T produce();
+
+}